refactor: rename
Rename modules: peaky_data.py -> data.py, strings.py -> formats.py, peaks_in_matplotlib.py -> interface_matplotlib.py, peaks_in_polars.py -> interface_polars.py, peaks_in_pandas.py -> interface_pandas.py, peaks_and_valleys.py -> peaks.py.
Rename classes: NumSlice -> Scope, SliceStr -> Region.
