+++ peakoscope/interface_matplotlib.py
